requirements.txt
adal==1.2.0
aniso8601==4.0.1
asn1crypto==0.24.0
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
Flask==1.0.2
Flask-RESTful==0.3.6
google-auth==1.6.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kubernetes==8.0.0
MarkupSafe==1.1.0
oauthlib==2.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.54.0
Werkzeug==0.14.1
